Share reagraph canvas; drop type colors; add icons
Replaces the bespoke Relationships-tab SVG with the same
reagraph canvas used on the projects page, including its
toolbar, zoom controls, and double-click-to-open
behaviour. Both views now render nodes as project_type
icons where defined (via getIconUrl) with neutral fill
elsewhere, and edges draw arrows toward the depended-on
project. The center project in the relationships graph
keeps the amber brand color so it stands out.
